Add an access token client to the mapper client
