+++ HappyTravel.Osaka.Api/Infrastructure/HttpClientNames.cs
@@ -3,5 +3,6 @@
     public static class HttpClientNames
     {
         public const string MapperApi = "Mapper";
+        public const string MapperIdentityClient = "MapperIdentityClient";
     }
 }
